Added use case and custom exception

+++ src/main/java/com/jmj/fibonacci/infrastructure/rest/FibonacciV1Controller.java
@@ -25,7 +25,7 @@ public class FibonacciV1Controller {
 
     @GetMapping("/calculate/{value}")
     public ResultV1DTO calculate(
-            @PathVariable(value = "value") Long number
+            @PathVariable(value = "value") int number
     ) {
         return longToResultV1DTOMapper.map(
                 calculateFibonacci.execute(new NumberToCalculate(number))
